remove plugin from rollup
