+++ css/src/tailwind.config.ts
@@ -9,7 +9,6 @@ import detailsOpenVariantPlugin from './tw-details-open-variant-plugin'
 import plugin from 'tailwindcss/plugin'
 import containerPlugin from '@tailwindcss/container-queries'
 import Hocus from 'tailwindcss-hocus'
-import componentPlugins from '@cypress-design/component-tailwind-plugins'
 
 function defineConfig(config: Config) {
   return config
@@ -72,7 +71,6 @@ export default (fileGlobs: string[] = []) => {
           },
         })
       }),
-      ...componentPlugins,
     ],
   })
 }
