--- packages/component-tailwind-plugins/ReadMe.md
@@ -1,7 +1,0 @@
-# @cypress-design/component-tailwind-plugins
-
-This package is the result by `scripts/generate-component-tailwind-plugins-package.mjs`
-
-⚠️ do not modify manually except the readme ⚠️
-
-It makes tailwind aware of the components in the design system that it will have to render.
