--- packages/component-tailwind-plugins/index.d.ts
@@ -1,3 +1,0 @@
-import plugin from 'tailwindcss/plugin'
-declare const listOfPugins: ReturnType<typeof plugin>[]
-export default listOfPugins
