--- packages/component-tailwind-plugins/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@cypress-design/component-tailwind-plugins",
-  "version": "0.0.0",
-  "private": "true",
-  "exports": {
-    "import": "./index.mjs",
-    "types": "./dist/index.d.ts"
-  },
-  "devDependencies": {
-    "@cypress-design/constants-accordion": "*",
-		"@cypress-design/constants-alert": "*",
-		"@cypress-design/constants-button": "*",
-		"@cypress-design/constants-checkbox": "*",
-		"@cypress-design/constants-docmenu": "*",
-		"@cypress-design/constants-menu": "*",
-		"@cypress-design/constants-modal": "*",
-		"@cypress-design/constants-spinner": "*",
-		"@cypress-design/constants-statusicon": "*",
-		"@cypress-design/constants-tabs": "*",
-		"@cypress-design/constants-tag": "*",
-		"@cypress-design/constants-testresult": "*",
-    "tailwindcss": "^3.4.3"
-  }
-}
